requirements.txt
Flask
Flask-SQLAlchemy
mysqlclient
python-dotenv
Werkzeug